api user and customer ok

+++ server/src/controllers/customer.controller.js
@@ -1,9 +1,16 @@
+import { Customer } from "../models/customer.model.js";
 import customerService from "../services/customer.service.js";
 
 class CustomerController {
   async getAllCustomers(req, res) {
     // transfer the logic to the service
     const users = await customerService.getAllCustomer();
+    if (!users || users.length === 0) {
+      return res.status(404).json({
+        success: false,
+        message: "No customers found",
+      });
+    }
     res.json({ success: true, data: users });
   }
 
@@ -24,7 +31,8 @@ class CustomerController {
 
   async updateCustomer(req, res) {
     try {
-      const { id } = req.user.id;
+      const id = req.user?.sub;
+      if (!id) return res.status(401).json({ success: false, message: "Unauthorized" });
       const data = req.body;
       // transfer the logic to the service
       const updateUser = await customerService.updateCustomer(id, data);
@@ -41,7 +49,8 @@ class CustomerController {
 
   async deleteCustomer(req, res, next) {
     try {
-      const { id } = req.user.id;
+      const id = req.user?.sub;
+      if (!id) return res.status(401).json({ success: false, message: "Unauthorized" });
       // transfer the logic to the service
       const result = await customerService.deleteCustomer(id);
       // return
@@ -49,6 +58,65 @@ class CustomerController {
         success: true,
         message: result.message,
       });
+    } catch (error) {
+      next(error);
+    }
+  }
+
+  async getAddresses(req, res, next) {
+    try {
+      const userId = req.user.sub;
+      const addressId = req.params;
+
+      const addresses = await customerService.getAddresses(userId, addressId);
+      res.status(200).json({ success: true, data: addresses });
+    } catch (error) {
+      next(error);
+    }
+  }
+
+  async getListAddresses(req, res, next) {
+    try {
+      const id = req.user.sub;
+      const addresses = await customerService.getListAddresses(id);
+      res.status(200).json({ success: true, data: addresses });
+    } catch (error) {
+      next(error);
+    }
+  }
+
+  async addAddress(req, res, next) {
+    try {
+      const id = req.user.sub;
+      const data = req.body;
+
+      const addresses = await customerService.addAddress(id, data);
+      res.status(200).json({ success: true, message: "Added new address", data: addresses });
+    } catch (error) {
+      next(error);
+    }
+  }
+
+  async updateAddress(req, res, next) {
+    try {
+      const addressId = req.params;
+      const userId = req.user.sub;
+      const data = req.body;
+
+      const address = await customerService.updateAddress(userId, addressId, data);
+      res.status(200).json({ success: true, message: "Address updated", data: address });
+    } catch (error) {
+      next(error);
+    }
+  }
+
+  async deleteAddress(req, res, next) {
+    try {
+      const userId = req.user.sub;
+      const addressId = req.params;
+
+      const addAddress = await customerService.deleteAddress(userId, addressId);
+      res.status(200).json({ success: true, message: "Address deleted", data: addAddress });
     } catch (error) {
       next(error);
     }
