+++ server/src/controllers/user.controller.js
@@ -3,9 +3,31 @@ import userService from "../services/user.service.js";
 
 class UserController {
   async getAllUsers(req, res) {
-    // transfer the logic to the service
-    const users = await userService.getAllUsers();
-    res.json({ success: true, data: users });
+    try {
+      const users = await userService.getAllUsers();
+      // if not found
+      if (!users || users.length === 0) {
+        return res.status(404).json({
+          success: false,
+          message: "No users found",
+        });
+      }
+      // if found -> return
+      res.status(200).json({ success: true, data: users });
+    } catch (error) {
+      next(error);
+    }
+  }
+
+  async getMyInfo(req, res) {
+    const myId = req.user?.sub;
+    if (!myId) return res.status(401).json({ success: false, message: "Unauthorized" });
+
+    const myInfo = await userService.getMyInfo(myId);
+    if (!myInfo) {
+      return res.status(404).json({ success: false, message: "Not found" });
+    }
+    res.json({ success: true, data: myInfo });
   }
 
   async getUserById(req, res) {
@@ -23,13 +45,38 @@ class UserController {
     res.json({ success: true, data: user });
   }
 
-  async updateUser(req, res) {
+  async updateMyInfo(req, res, next) {
     try {
-      const { id } = req.user.id;
+      // const id = req.user.sub;
+      const id = req.user?.sub;
+      if (!id) return res.status(401).json({ success: false, message: "Unauthorized" });
       const data = req.body;
       // transfer the logic to the service
-      const updateUser = await userService.updateUser(id, data);
+      const updateUser = await userService.updateMyInfo(id, data);
       // return
+      return res.status(200).json({
+        success: true,
+        message: "Information updated successfully.",
+        data: updateUser,
+      });
+    } catch (error) {
+      next(error);
+    }
+  }
+
+  async updateUser(req, res, next) {
+    try {
+      // check role
+      const role = req.user?.role;
+      if (role !== "admin") {
+        return res.status(401).json({ success: false, message: "Unauthorized" });
+      }
+
+      const data = req.body;
+      const { id } = req.params.id;
+      // transfer the logic to the service
+      const updateUser = await userService.updateUser(id, data);
+
       return res.status(200).json({
         success: true,
         message: "Information updated successfully.",
@@ -42,7 +89,8 @@ class UserController {
 
   async deleteUser(req, res, next) {
     try {
-      const { id } = req.user.id;
+      const id = req.user?.sub;
+      if (!id) return res.status(401).json({ success: false, message: "Unauthorized" });
       // transfer the logic to the service
       const result = await userService.deleteUser(id);
       // delete cookie token session
